+++ deepfellow/infra/model/__init__.py
@@ -12,10 +12,12 @@
 import typer
 
 from .install import app as install_app
+from .list import app as list_app
 from .uninstall import app as uninstall_app
 
 app = typer.Typer(no_args_is_help=True)
 
 
 app.add_typer(install_app)
+app.add_typer(list_app)
 app.add_typer(uninstall_app)